drivers: add param_read_write fake
@nathanchance: squashed and cleaned up these commits in Erik's tree.
               Wrote commit message as well.

https://github.com/pappschlumpf/op6/commit/0aa2f121843692103e758bc1d7a61c66148db219
https://github.com/pappschlumpf/op6/commit/5da504bde8abb0cf1ea3641031a5e8c9f33cfc6a
https://github.com/pappschlumpf/op6/commit/d032fe2991e2fa7744372d10676382bb8c6bc6fe

We don't need a dedicated Kconfig (because we always want this enabled)
or Makefile (because it's just one file).

OnePlus did not include this driver in their initial check-in (which is
a whole other thing with the GPL and such...) so when users flash a
custom kernel, the paths for this driver disappear. This affects users
who have an Avengers OP6 as the cust_flag path tells OxygenOS they have
that type of device and they can no longer select the theme (even worse,
if the theme was selected and they want to change it, their Settings
will crash due to an array out of bounds exception).

Normally, the sysfs path will display the value stored in the param
partition. Rather than trying reverse engineer this driver based on the
OP5 so we can read/write to the partition directly, we can just create
the path with this super simple driver which results in the same thing
since userspace checks the value based on this path, not from the param
partition directly.

Some further info:

https://forum.xda-developers.com/showthread.php?p=77024052
https://forum.xda-developers.com/showthread.php?p=77032128

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>