[REAPPLY] cpumask: Add cpumasks for big and little CPU clusters
Add cpu_lp_mask and cpu_perf_mask to represent the CPUs that belong to
each cluster in a dual-cluster, heterogeneous system.

Note that this assumes that CONFIG_NR_CPUS is set to exatly the number of
CPUs present, and that the number of CPUs per cluster is split evenly in
half.

[@mcdachpappe] reapply partially these changes due to reverted irq-critical
api commit which were oddly applied with this commit 22f2fd6e9201667814f26df0276dcebf9f9bd680
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Signed-off-by: mcdachpappe <noreference@web.de>

Conflicts:
	kernel/cpu.c