defconfig: Disable SWAP/ZRAM and dependancies
After a fairly lengthy trial run, I'm making an executive decision
to shut this shit down. Neutrino has run without it for 3/4 of it's
lifespan and I wasn't hearing any complaints then. My initial
intention here was to implement a toned-down variation of what
OnePlus had decided to do with their stock kernel, and then I
remembered why distancing my codebase as far away as possible from
OP's was one of my primary objectives with this project.

Throwing ~2GB of zRAM (not to mention an additional 2GB of
swapspace) at a device with 6-8GB of onboard RAM is an absurd
bandaid that at best only serves to delay the inevitable and
likely won't even do that. OnePlus needs to get their task
management in check rather than trying to hide their shortcomings
behind a wall of virtual memory and that's the bottom line.

Furthermore, this is the only relevant aspect of the kernel that
I am still having to rely on Magisk to configure properly and
there is nothing more I'd rather do than to cut that tether
once and for all, so, consider it done.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>

# Conflicts:
#	arch/arm64/configs/radioactive_defconfig